numeric: Move iota from algorithm header
