+++ src/stdgpu/impl/algorithm_detail.h
@@ -60,39 +60,6 @@ for_each_index(ExecutionPolicy&& policy, IndexType size, UnaryFunction f)
 namespace detail
 {
 template <typename Iterator, typename T>
-class iota_functor
-{
-public:
-    iota_functor(Iterator begin, T value)
-      : _begin(begin)
-      , _value(value)
-    {
-    }
-
-    STDGPU_HOST_DEVICE void
-    operator()(const index_t i)
-    {
-        _begin[i] = _value + static_cast<T>(i);
-    }
-
-private:
-    Iterator _begin;
-    T _value;
-};
-} // namespace detail
-
-template <typename ExecutionPolicy, typename Iterator, typename T>
-void
-iota(ExecutionPolicy&& policy, Iterator begin, Iterator end, T value)
-{
-    for_each_index(std::forward<ExecutionPolicy>(policy),
-                   static_cast<index_t>(end - begin),
-                   detail::iota_functor<Iterator, T>(begin, value));
-}
-
-namespace detail
-{
-template <typename Iterator, typename T>
 class fill_functor
 {
 public:
